+++ auto_route_generator/lib/src/builders/auto_router_builder.dart
@@ -47,11 +47,15 @@ class AutoRouterBuilder extends CacheAwareBuilder<RouterConfig> {
     for (final clazz in unit.declarations
         .whereType<ClassDeclaration>()
         .where((e) => e.metadata.any((e) => e.name.name == annotationName))) {
-      final routerAnnotation = clazz.metadata.firstWhere((e) => e.name.name == annotationName);
-      final partDirectives =
-          unit.directives.whereType<PartDirective>().fold<int>(0, (acc, a) => acc ^ a.toSource().hashCode);
-      calculatedHash =
-          calculatedHash ^ clazz.name.toString().hashCode ^ routerAnnotation.toSource().hashCode ^ partDirectives;
+      final routerAnnotation =
+          clazz.metadata.firstWhere((e) => e.name.name == annotationName);
+      final partDirectives = unit.directives
+          .whereType<PartDirective>()
+          .fold<int>(0, (acc, a) => acc ^ a.toSource().hashCode);
+      calculatedHash = calculatedHash ^
+          clazz.name.toString().hashCode ^
+          routerAnnotation.toSource().hashCode ^
+          partDirectives;
     }
     return calculatedHash;
   }
@@ -65,7 +69,8 @@ class AutoRouterBuilder extends CacheAwareBuilder<RouterConfig> {
   }
 
   @override
-  Future<String> onGenerateContent(BuildStep buildStep, RouterConfig item) async {
+  Future<String> onGenerateContent(
+      BuildStep buildStep, RouterConfig item) async {
     final generateForDir = item.generateForDir;
     final generatedResults = <RoutesList>[];
     final routes = <RouteConfig>[];
@@ -96,7 +101,8 @@ class AutoRouterBuilder extends CacheAwareBuilder<RouterConfig> {
   }
 
   @override
-  Future<RouterConfig?> onResolver(LibraryReader library, BuildStep buildStep, int stepHash) async {
+  Future<RouterConfig?> onResolver(
+      LibraryReader library, BuildStep buildStep, int stepHash) async {
     final annotatedElements = library.annotatedWith(_typeChecker);
     if (annotatedElements.isEmpty) return null;
     final element = annotatedElements.first.element;
@@ -121,7 +127,8 @@ class AutoRouterBuilder extends CacheAwareBuilder<RouterConfig> {
     return router;
   }
 
-  void _writeRouterFile(BuildStep buildStep, RouterConfig router, {bool toCache = false}) {
+  void _writeRouterFile(BuildStep buildStep, RouterConfig router,
+      {bool toCache = false}) {
     final routerFile = _buildRouterFile(buildStep, toCache: toCache);
     if (!routerFile.existsSync()) {
       routerFile.createSync(recursive: true);